render json, except for notebooks
